feat(signatures): Fixed `signAmino` & replaced lum deps
