+++ jestconfig.json
@@ -2,6 +2,6 @@
   "transform": {
     "^.+\\.(t|j)sx?$": "ts-jest"
   },
-  "testRegex": "/test/.*\\.test\\.ts$",
+  "testRegex": "/tests/.*\\.test\\.ts$",
   "collectCoverageFrom": ["src/**/*.{ts,js}"]
 }
