+++ package.json
@@ -58,8 +58,7 @@
     "@cosmjs/amino": "^0.28.4",
     "@cosmjs/encoding": "^0.28.4",
     "@cosmjs/proto-signing": "^0.28.4",
-    "@cosmjs/stargate": "^0.28.13",
-    "@lum-network/sdk-javascript": "^0.7.6",
+    "@cosmjs/stargate": "^0.28.4",
     "browserify-fs": "^1.0.0",
     "buffer": "6.0.3",
     "cosmjs-types": "^0.5.0",
@@ -70,7 +69,6 @@
     "path-browserify": "1.0.1",
     "process": "^0.11.10",
     "stream-browserify": "3.0.0",
-    "tiny-secp256k1": "^2.2.1",
     "util": "^0.12.4"
   },
   "publishConfig": {
